requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\Koteusz\Desktop\discordBot\giveaway\main_giveaway.py: 1,2,3
# C:\Users\Koteusz\Desktop\discordBot\main.py: 1,5
# C:\Users\Koteusz\Desktop\discordBot\messages\main_messages.py: 1,2,3
# C:\Users\Koteusz\Desktop\discordBot\stream_notifications\main_stream_notifications.py: 1,2,3
# C:\Users\Koteusz\Desktop\discordBot\user_managment\main_user_managment.py: 1,2,4
discord.py == 2.0.0a

# C:\Users\Koteusz\Desktop\discordBot\stay_alive.py: 1
flask == 2.0.1

# C:\Users\Koteusz\Desktop\discordBot\giveaway\main_giveaway.py: 9
python_dateutil == 2.8.1

# C:\Users\Koteusz\Desktop\discordBot\main.py: 6
python_dotenv == 0.18.0
